fix: Fix attribute error when deleting emails (#23845)
We just missed a guard here

Fixes SENTRY-NBA

+++ src/sentry/api/endpoints/user_emails.py
@@ -215,7 +215,8 @@ class UserEmailsEndpoint(UserEndpoint):
         if primary_email == del_email:
             return self.respond({"detail": "Cannot remove primary email"}, status=400)
 
-        del_email.delete()
+        if del_email:
+            del_email.delete()
 
         logger.info(
             "user.email.remove",
